requirements.txt
discord.py
python-dotenv
python-a2s